postRequest, deleted console logs
